+++ app/src/main/java/com/example/fuzzynumbercomparator/fuzzynumber/FuzzyMaxOrderComparingStrategy.kt
@@ -4,19 +4,19 @@ package com.example.fuzzynumbercomparator.fuzzynumber
 class FuzzyMaxOrderComparingStrategy : ComparingStrategy {
     private val subAreas = SubAreaResults()
 
-    override fun compare(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Float {
+    override fun compare(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Double {
         subAreas.setDefaultValues()
         return when {
-            firstNumber.getStart() > secondNumber.getEnd() -> 1f
-            secondNumber.getStart() > firstNumber.getEnd() -> 0f
+            firstNumber.getStart() > secondNumber.getEnd() -> 1.0
+            secondNumber.getStart() > firstNumber.getEnd() -> 0.0
             else -> calculateComparison(firstNumber, secondNumber)
         }
     }
 
-    private fun calculateComparison(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Float {
+    private fun calculateComparison(firstNumber: FuzzyNumber, secondNumber: FuzzyNumber): Double {
         val coordinates = LinearAlgebra.getIntersectionCoordinatesAndRange(firstNumber, secondNumber)
         for(i in coordinates.size - 1 downTo 1) {
-            val middleCoordinate = (coordinates[i] + coordinates[i - 1]) / 2f
+            val middleCoordinate = (coordinates[i] + coordinates[i - 1]) / 2.0
             if(LinearAlgebra.areCoordinatesOnVerticalLine(middleCoordinate, coordinates[i]))
                 continue
 
@@ -36,23 +36,23 @@ class FuzzyMaxOrderComparingStrategy : ComparingStrategy {
     }
 
     private fun updateSubArea(
-        firstValue: Float,
-        secondValue: Float,
-        x1: Float,
-        x2: Float,
+        firstValue: Double,
+        secondValue: Double,
+        x1: Double,
+        x2: Double,
         firstNumber: FuzzyNumber,
         secondNumber: FuzzyNumber) {
         when {
-            firstValue == 0f && firstNumber.isBiggerThanValue(x1) -> {
+            firstValue == 0.0 && firstNumber.isBiggerThanValue(x1) -> {
                 subAreas.b += secondNumber.subArea(x1, x2)
             }
-            firstValue == 0f && firstNumber.isSmallerThanValue(x2) -> {
+            firstValue == 0.0 && firstNumber.isSmallerThanValue(x2) -> {
                 subAreas.c += secondNumber.subArea(x1, x2)
             }
-            secondValue == 0f && secondNumber.isBiggerThanValue(x1)-> {
+            secondValue == 0.0 && secondNumber.isBiggerThanValue(x1)-> {
                 subAreas.c += firstNumber.subArea(x1, x2)
             }
-            secondValue == 0f && secondNumber.isSmallerThanValue(x2) -> {
+            secondValue == 0.0 && secondNumber.isSmallerThanValue(x2) -> {
                 subAreas.b += firstNumber.subArea(x1, x2)
             }
             firstValue > secondValue -> {
@@ -64,24 +64,24 @@ class FuzzyMaxOrderComparingStrategy : ComparingStrategy {
         }
     }
 
-    private fun calculateResult(): Float {
+    private fun calculateResult(): Double {
         val result = (subAreas.d + subAreas.b) / subAreas.totalSubArea()
-        return if (result.isNaN()) 0.5f else result
+        return if (result.isNaN()) 0.5 else result
     }
 }
 
 data class SubAreaResults(
-    var a: Float = 0f,
-    var b: Float = 0f,
-    var c: Float = 0f,
-    var d: Float = 0f,
+    var a: Double = 0.0,
+    var b: Double = 0.0,
+    var c: Double = 0.0,
+    var d: Double = 0.0,
 ) {
     fun totalSubArea() = a + b + c + d
     fun setDefaultValues() {
-        a = 0f
-        b = 0f
-        c = 0f
-        d = 0f
+        a = 0.0
+        b = 0.0
+        c = 0.0
+        d = 0.0
     }
 }
