+++ app/src/main/java/com/example/fuzzynumbercomparator/fuzzynumber/FuzzyNumberFactory.kt
@@ -1,5 +1,5 @@
 package com.example.fuzzynumbercomparator.fuzzynumber
 
 interface FuzzyNumberFactory {
-    fun createFromFuzzyNumberTypeWithParams(type: FuzzyNumberType, params: List<Float>): FuzzyNumber
+    fun createFromFuzzyNumberTypeWithParams(type: FuzzyNumberType, params: List<Double>): FuzzyNumber
 }
